requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
httpx>=0.27.2
python-dotenv>=1.0.1
loguru>=0.7.2
